# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PinIt Security & Antivirus Detection API
  description: |
    Enterprise-grade security analysis API for scanning uploaded files and URLs for malware,
    viruses, phishing, macros, obfuscated scripts, and suspicious behavior.
  version: 1.0.0
  contact:
    name: Security Engineering Team
    url: https://pinit.security
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
paths:
...
  /analyze/file:
    post:
      summary: Upload and securely analyze file
      description: Safely inspects files in quarantine storage with static multi-engine analysis (never executes code). Supports synchronous analysis (?sync=true) returning file_type, file_size, detected_indicators, risk_score, classification, evidence, and recommended_action.
      parameters:
        - in: query
          name: sync
          schema:
            type: boolean
          description: When true, executes analysis synchronously and returns completed results immediately.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Synchronous analysis result with exact 7 security fields.
        '202':
          description: Queued analysis response.
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/FileTooLarge'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Invalid request parameters or payload.
...
    FileTooLarge:
      description: Uploaded file exceeds size limit.